Handle selecting and pasting
I had considered pasting _new_ chars
I had considered selecting and deleting chars
I had not considered selecting and pasting (effectively replacing chars)

It should now handle that.

Only thing it won't do is if you select and paste the same char as the
mask char it will not replace, but that is an edge case I am willing
to live with.